+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # Changelog
 
+## [Unreleased]
+
+## [1.2.0]
+
+### Added
+
+- Add basic support for rocksolid custom elements
+- Add configuration option to adjust supported content elements and frontend modules
+
 ## [1.1.0]
 
 ### Added
@@ -8,4 +17,5 @@
 - Add info messages if no consent tag is assign for supported frontend modules and content elements
 
 
+[1.2.0]: https://github.com/hofff/contao-consent-bridge/compare/1.0.0...1.1.0
 [1.1.0]: https://github.com/hofff/contao-consent-bridge/compare/1.0.0...1.1.0